pyproject.toml
[project]
name = "autodev-agent"
version = "0.1.0"
description = "Autonomous development improvement agent for Privacy by Design Foundation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Privacy by Design Foundation"}
]
keywords = ["automation", "ci/cd", "llm", "privacy", "digital-identity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "markdownify>=0.12.0",
    "PyGithub>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "responses>=0.25.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
autodev = "autodev.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
